requirements.txt
pyyaml>=6.0
pydantic>=2.0
feedparser>=6.0
pytz>=2023.3
requests>=2.31.0
